Nine-Box Grid
A “nine-box grid” is a matrix tool that is used to evaluate and plot a
company’s talent pool, based on two factors, which most commonly are
performance and potential.  Typically on the horizontal axis is
"performance" - measured by performance reviews.  On the vertical axis
is typically "potential" - referring to an individual's potential to
grow one or more levels in a managerial or professional capacity.